refactor: StaffLayOut 을 작성해 Vue Router 구조가 어느정도 잡힌 상황에 맞춰 Staff 관련 Vue 들 components, layouts, views 들로 구분해 디렉토리 구조 리팩토링 수행.
