+++ frontend/src/reception/components/WaitingList.vue
@@ -4,9 +4,9 @@
   import WaitingListDoctorName from "@/shared/components/WaitingListDoctorName.vue";
   import WaitingListPatientList from "@/shared/components/WaitingListPatientList.vue";
   import {reception} from "@/reception/util/reception.js";
-import { getStompClient, sendMsg, subscribeChannel } from "@/util/stompMethod";
-import { useUserStore } from "@/stores/userStore";
-import { getAccessToken } from "@/auth/accessToken";
+import { getStompClient, sendMsg, subscribeChannel } from "@/util/stompMethod.js";
+import { useUserStore } from "@/stores/userStore.js";
+import { getAccessToken } from "@/auth/accessToken.js";
 import { createWebSocketModuleRunnerTransport } from "vite/module-runner";
 import VisitHistory from "@/components/diagnosis/VisitHistory.vue";
   import {useRouter} from "vue-router";
