+++ frontend/src/router/index.js
@@ -3,7 +3,7 @@ import { patientRoutes } from './patientRoutes';
 import { adminRoutes } from './adminRoutes';
 import { staffRoutes } from './staffRoutes';
 
-const StaffLayOut = () => import('@/staff/views/StaffLayOut.vue')
+const StaffLayOut = () => import('@/staff/layouts/StaffLayOut.vue')
 
 const router = createRouter({
   history: createWebHistory(import.meta.env.BASE_URL),
