+++ frontend/src/router/patientRoutes.js
@@ -4,8 +4,8 @@ const PatientLayout = () => import('@/patient/layouts/PatientLayout.vue');
 const RegisterView = () => import("@/patient/views/RegisterView.vue");
 const UpdateProfile = () => import("@/patient/views/UpdateProfile.vue");
 const LoginView = () => import('@/patient/views/LoginView.vue');
-const AcceptPatientByStaff = () => import("@/reception/views/AcceptPatientByStaff.vue");
-const RegReservation = () => import('@/reservation/views/RegReservation.vue')
+const AcceptPatientByStaff = () => import("@/staff/views/AcceptPatientByStaff.vue");
+const RegReservation = () => import('@/shared/views/RegReservation.vue')
 const ReservationListByPatient = () => import('@/reservation/views/ReservationListByPatient.vue')
 
 export const patientRoutes = [
