+++ frontend/src/shared/views/RegReservation.vue
