+++ frontend/src/staff/components/WaitingReservationParent.vue
@@ -2,8 +2,8 @@
 import '@/css/staff/waitingreservaion.css'
 import {useRouter} from "vue-router";
 import {reactive} from "vue";
-import ReservationListByStaff from "@/reservation/views/ReservationListByStaff.vue";
-import WaitingList from "@/reception/views/WaitingList.vue";
+import ReservationListByStaff from "@/staff/views/ReservationListByStaff.vue";
+import WaitingList from "@/reception/components/WaitingList.vue";
 
 const router = useRouter();
