+++ frontend/src/staff/views/AcceptPatientByStaff.vue
