+++ frontend/src/staff/views/RegPatientByStaff.vue
